requirements.txt
faiss-cpu==1.8.0.post1
ftfy==6.2.3
googletrans==3.1.0a0
huggingface-hub==0.25.1
nltk==3.9.1
numpy==1.26.4
open_clip_torch==2.26.1
pandas==2.2.3
pillow==10.2.0
streamlit-nightly==1.38.1.dev20240925
tqdm==4.66.5

